Code block: move parsing functions to a separate module
This reduces clutter in the cog. The cog should only have Discord-
related functionality.
